+++ test/unit/index.js
@@ -21,6 +21,7 @@ require('./spec/Supervisor');
 require('./spec/Task');
 require('./spec/Worker');
 require('./spec/WorkerEvents');
+require('./spec/Workspace');
 require('./spec/signaling/EventBridgeSignaling');
 require('./spec/util/Configuration');
 require('./spec/util/Logger');